logtestutils: extract telemetry stubs to break import cycle
The logtestutils package contained telemetry test stubs (StubTime,
StubQueryStats, StubTracingStatus) that import pkg/sql/execstats.
This created a dependency cycle when pkg/obs/ash tests imported
logtestutils for the log spy utilities:

  ash_test → logtestutils → execstats → ... → admission → ash

Move the telemetry stubs into a new sub-package
logtestutils/telemetrylogtestutils so that importing logtestutils
for LogSpy/StructuredLogSpy no longer transitively pulls in
execstats.

Epic: CRDB-44217

Release note: None
