+++ pkg/util/log/logtestutils/BUILD.bazel
@@ -6,12 +6,10 @@ go_library(
         "log_spy.go",
         "log_test_utils.go",
         "structured_log_spy.go",
-        "telemetry_logging_test_utils.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/util/log/logtestutils",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/sql/execstats",
         "//pkg/util/log",
         "//pkg/util/log/logconfig",
         "//pkg/util/log/logpb",
