+++ pkg/util/log/logtestutils/telemetrylogtestutils/telemetry_logging_test_utils.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package logtestutils
+package telemetrylogtestutils
 
 import (
 	"time"
